added last_active variable to Characters for more accurate association of procs to powers when pets are present.
